exception for products are visible

+++ src/main/java/com/uade/tpo/marketplace/controllers/products/ProductRequest.java
@@ -10,7 +10,7 @@ public class ProductRequest {
     protected String title;
     protected String author;
     protected String description;
-    protected float price;
+    protected double price;
     protected int stock;
     protected List<String> urlImage;
