+++ src/main/java/com/uade/tpo/marketplace/service/MusicAlbumService.java
@@ -23,7 +23,7 @@ public interface MusicAlbumService {
 
     public Page<MusicAlbum> getMusicAlbumByAuthor(String author, PageRequest pageable);
 
-    public MusicAlbum createMusicAlbum(String title, String author, String recordLabel,int year, String description, String isrc, List<Genre> genres, Float price,
+    public MusicAlbum createMusicAlbum(String title, String author, String recordLabel,int year, String description, String isrc, List<Genre> genres, double price,
     
     List<String> urlImages) throws MusicAlbumDuplicateException;
